+++ Core/build.gradle
@@ -16,7 +16,7 @@ dependencies {
     compile group: 'com.google.guava', name: 'guava', version: '17.0'
     compile group: 'com.cedarsoftware', name: 'json-io', version: '4.12.0'
     compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.0'
-    compile group: 'com.github.Vatuu', name: 'discord-rpc', version: '1.6.2'
+    compile group: 'com.jagrosh', name: 'DiscordIPC', version: '0.4'
 }
 
 jar {
