+++ Core/src/main/java/org/sorus/client/gui/screen/settings/SettingsScreen.java
@@ -24,217 +24,13 @@
 
 package org.sorus.client.gui.screen.settings;
 
-import java.awt.*;
 import org.sorus.client.Sorus;
-import org.sorus.client.gui.core.Screen;
-import org.sorus.client.gui.core.component.Collection;
-import org.sorus.client.gui.core.component.Panel;
-import org.sorus.client.gui.core.component.impl.*;
-import org.sorus.client.gui.core.component.impl.Rectangle;
-import org.sorus.client.gui.core.font.IFontRenderer;
-import org.sorus.client.util.MathUtil;
-import org.sorus.client.version.input.Key;
+import org.sorus.client.gui.core.ThemeableScreen;
 
-/** Screen used with {@link IConfigurableScreen} to display and allow configuration of settings. */
-public class SettingsScreen extends Screen {
-
-  private final IConfigurableScreen configurable;
-
-  private Panel main;
-  private Scroll scroll;
-  private SettingsHolder settingsHolder;
-
-  private double targetScroll;
+public class SettingsScreen extends ThemeableScreen {
 
   public SettingsScreen(IConfigurableScreen configurable) {
-    this.configurable = configurable;
-  }
-
-  @Override
-  public void init() {
-    Sorus.getSorus().getVersion().getRenderer().enableBlur();
-    main = new Panel();
-    Collection menu = new Collection().position(610, 140);
-    main.add(menu);
-    menu.add(new Rectangle().smooth(5).size(700, 720).position(0, 70).color(new Color(18, 18, 18)));
-    menu.add(new Rectangle().size(700, 65).position(0, 5).color(new Color(30, 30, 30)));
-    menu.add(new Arc().radius(5, 5).angle(180, 270).position(0, 0).color(new Color(30, 30, 30)));
-    menu.add(new Arc().radius(5, 5).angle(90, 180).position(690, 0).color(new Color(30, 30, 30)));
-    menu.add(new Rectangle().size(690, 5).position(5, 0).color(new Color(30, 30, 30)));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14),
-                new Color(14, 14, 14))
-            .size(700, 7)
-            .position(0, 70));
-    IFontRenderer fontRenderer =
-        Sorus.getSorus().getGUIManager().getRenderer().getRubikFontRenderer();
-    menu.add(
-        new Text()
-            .fontRenderer(fontRenderer)
-            .text("SORUS")
-            .position(350 - fontRenderer.getStringWidth("SORUS") / 2 * 5.5, 17.5)
-            .scale(5.5, 5.5)
-            .color(new Color(215, 215, 215)));
-    menu.add(
-        new Text()
-            .fontRenderer(fontRenderer)
-            .text(configurable.getDisplayName())
-            .position(
-                350 - fontRenderer.getStringWidth(configurable.getDisplayName()) / 2 * 4.5, 95)
-            .scale(4.5, 4.5)
-            .color(new Color(175, 175, 175)));
-    menu.add(new Rectangle().size(660, 620).position(20, 150).color(new Color(30, 30, 30)));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30))
-            .size(660, 4)
-            .position(20, 146));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30))
-            .size(4, 4)
-            .position(680, 146));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 150))
-            .size(4, 620)
-            .position(680, 150));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 150))
-            .size(4, 4)
-            .position(680, 770));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 150))
-            .size(660, 4)
-            .position(20, 770));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 30))
-            .size(4, 4)
-            .position(16, 770));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 30))
-            .size(4, 620)
-            .position(16, 150));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 150),
-                new Color(14, 14, 14, 30),
-                new Color(14, 14, 14, 30))
-            .size(4, 4)
-            .position(16, 146));
-    /*menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14),
-                new Color(14, 14, 14))
-            .size(650, 7)
-            .position(25, 770));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14),
-                new Color(14, 14, 14),
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14, 0))
-            .size(650, 7)
-            .position(25, 143));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14),
-                new Color(14, 14, 14),
-                new Color(14, 14, 14, 0))
-            .size(7, 610)
-            .position(13, 155));
-    menu.add(
-        new Rectangle()
-            .gradient(
-                new Color(14, 14, 14),
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14, 0),
-                new Color(14, 14, 14))
-            .size(7, 610)
-            .position(680, 155));*/
-    Scissor scissor = new Scissor().size(680, 580).position(10, 170);
-    this.scroll = new Scroll();
-    scroll.position(0, 2);
-    scissor.add(scroll);
-    menu.add(scissor);
-    scissor.add(scroll);
-    settingsHolder = new SettingsHolder();
-    configurable.addConfigComponents(settingsHolder);
-    scroll.add(settingsHolder);
-    menu.add(scissor);
-  }
-
-  @Override
-  public void onRender() {
-    main.scale(
-        Sorus.getSorus().getVersion().getScreen().getScaledWidth() / 1920,
-        Sorus.getSorus().getVersion().getScreen().getScaledHeight() / 1080);
-    final int FPS = Math.max(Sorus.getSorus().getVersion().getGame().getFPS(), 1);
-    double scrollValue = Sorus.getSorus().getVersion().getInput().getScroll();
-    targetScroll = targetScroll + scrollValue * 0.7;
-    scroll.setScroll((targetScroll - scroll.getScroll()) * 7 / FPS + scroll.getScroll());
-    double maxScroll = settingsHolder.getHeight() - 580;
-    scroll.addMinMaxScroll(-maxScroll, 0);
-    targetScroll = MathUtil.clamp(targetScroll, scroll.getMinScroll(), scroll.getMaxScroll());
-    main.onRender(this);
-  }
-
-  @Override
-  public void exit() {
-    Sorus.getSorus().getVersion().getRenderer().disableBlur();
-    Sorus.getSorus().getSettingsManager().save();
-    main.onRemove();
-  }
-
-  @Override
-  public void keyTyped(Key key) {
-    if (key == Key.ESCAPE && this.isInteractContainer()) {
-      Sorus.getSorus().getGUIManager().close(this);
-    }
+    super(Sorus.getSorus().getThemeManager().getTheme("settings", configurable));
   }
 
   @Override
