+++ Core/src/main/java/org/sorus/client/gui/theme/defaultTheme/menu/MenuComponent.java
@@ -1,0 +1,167 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020 Danterus
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package org.sorus.client.gui.theme.defaultTheme.menu;
+
+import org.sorus.client.Sorus;
+import org.sorus.client.event.EventInvoked;
+import org.sorus.client.event.impl.client.input.MousePressEvent;
+import org.sorus.client.gui.core.Screen;
+import org.sorus.client.gui.core.component.Collection;
+import org.sorus.client.gui.core.component.impl.Rectangle;
+import org.sorus.client.gui.core.component.impl.Text;
+import org.sorus.client.util.MathUtil;
+
+import java.awt.*;
+
+public class MenuComponent extends Collection {
+
+  private final Collection main;
+  private final Text text;
+  private final Collection symbol;
+  private final Class<? extends Screen> screenClass;
+
+  private double hoveredPercent;
+
+  public MenuComponent(String name, ILogoCreator logoCreator, Class<? extends Screen> screenClass) {
+    this.screenClass = screenClass;
+    this.add(main = new Collection());
+    main.add(new Rectangle().size(167.5, 172.5).color(new Color(30, 30, 30)));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30))
+                    .size(167.5, 4)
+                    .position(0, -4));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30))
+                    .size(4, 4)
+                    .position(167.5, -4));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 150))
+                    .size(4, 172.5)
+                    .position(167.5, -4));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 150))
+                    .size(4, 4)
+                    .position(167.5, 172.5));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 150))
+                    .size(167.5, 4)
+                    .position(0, 172.5));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 30))
+                    .size(4, 4)
+                    .position(-4, 172.5));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 30))
+                    .size(4, 172.5)
+                    .position(-4, 0));
+    main.add(
+            new Rectangle()
+                    .gradient(
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 150),
+                            new Color(14, 14, 14, 30),
+                            new Color(14, 14, 14, 30))
+                    .size(4, 4)
+                    .position(-4, -4));
+    this.text = new Text()
+            .fontRenderer(Sorus.getSorus().getGUIManager().getRenderer().getGidoleFontRenderer()).text(name);
+    main.add(text.scale(3, 3).position(83.75 - text.width() * 3 / 2, 145));
+    main.add(symbol = new Collection());
+    symbol.add(new Text().fontRenderer(Sorus.getSorus().getGUIManager().getRenderer().getGidoleFontRenderer()).text("?").scale(12.5, 12.5).position(60, 10));
+    Sorus.getSorus().getEventManager().register(this);
+  }
+
+  @Override
+  public void onRender() {
+    boolean hovered = this.isHovered(Sorus.getSorus().getVersion().getInput().getMouseX(), Sorus.getSorus().getVersion().getInput().getMouseY());
+    hoveredPercent = MathUtil.clamp(hoveredPercent + (hovered ? 1 : -1) * 0.04, 0, 1);
+    main.position(-5 * hoveredPercent, -5 * hoveredPercent);
+    main.scale(1 + hoveredPercent * 0.05, 1 + hoveredPercent * 0.05);
+    int color = (int) (215 + 40 * hoveredPercent);
+    text.color(new Color(color, color, color));
+    symbol.color(new Color(color, color, color));
+    super.onRender();
+  }
+
+  @Override
+  public void onRemove() {
+    Sorus.getSorus().getEventManager().unregister(this);
+  }
+
+  @EventInvoked
+  public void onClick(MousePressEvent e) {
+    if(this.isHovered(e.getX(), e.getY())) {
+      Sorus.getSorus().getGUIManager().close((Screen) this.getContainer());
+      try {
+        Sorus.getSorus().getGUIManager().open(screenClass.newInstance());
+      } catch(InstantiationException | IllegalAccessException ex) {
+        ex.printStackTrace();
+      }
+    }
+  }
+
+  private boolean isHovered(double x, double y) {
+    return x > this.absoluteX() && x < this.absoluteX() + 167.5 * this.absoluteXScale() && y > this.absoluteY() && y < this.absoluteY() + 172.5 * this.absoluteYScale();
+  }
+
+  public interface ILogoCreator {
+    void addLogoComponents();
+  }
+}
