+++ Core/src/main/java/org/sorus/client/module/impl/discordrp/DiscordRP.java
@@ -1,19 +1,19 @@
 package org.sorus.client.module.impl.discordrp;
 
-import net.arikia.dev.drpc.DiscordEventHandlers;
-import net.arikia.dev.drpc.DiscordRPC;
-import net.arikia.dev.drpc.DiscordRichPresence;
+import com.jagrosh.discordipc.IPCClient;
+import com.jagrosh.discordipc.IPCListener;
+import com.jagrosh.discordipc.entities.RichPresence;
+import com.jagrosh.discordipc.exceptions.NoDiscordClientException;
+import java.time.OffsetDateTime;
 import org.sorus.client.Sorus;
 import org.sorus.client.event.EventInvoked;
 import org.sorus.client.event.impl.client.GuiSwitchEvent;
-import org.sorus.client.event.impl.client.StartEvent;
-import org.sorus.client.event.impl.client.TickEvent;
 import org.sorus.client.module.ModuleConfigurable;
 
 public class DiscordRP extends ModuleConfigurable {
 
-  private final DiscordRichPresence presence = new DiscordRichPresence();
-  private DiscordEventHandlers handlers;
+  private final IPCClient client = new IPCClient(763060164126834729L);
+  private final RichPresence.Builder builder = new RichPresence.Builder();
 
   public DiscordRP() {
     super("DISCORD RPC");
@@ -22,55 +22,43 @@ public class DiscordRP extends ModuleConfigurable {
 
   @Override
   public void onLoad() {
-    handlers = new DiscordEventHandlers.Builder().build();
-    presence.startTimestamp = System.currentTimeMillis();
-    presence.largeImageKey = "main";
-    Runtime.getRuntime().addShutdownHook(new Thread(DiscordRPC::discordShutdown));
+    client.setListener(new IPCListener() {});
+
+    builder.setStartTimestamp(OffsetDateTime.now()).setLargeImage("main", "Sorus Client");
+    Runtime.getRuntime().addShutdownHook(new Thread(client::close));
   }
 
   @Override
   public void onEnable() {
-    DiscordRPC.discordInitialize("763060164126834729", handlers, true);
+    try {
+      client.connect();
+      client.sendRichPresence(builder.build());
+    } catch (NoDiscordClientException e) {
+      e.printStackTrace();
+    }
   }
 
   @Override
   public void onDisable() {
-    DiscordRPC.discordShutdown();
-  }
-
-  @EventInvoked
-  public void onStart(StartEvent e) {
-    if (this.isEnabled()) {
-      this.onEnable();
-    }
-  }
-
-  @EventInvoked
-  public void onTick(TickEvent e) {
-    if (this.isEnabled()) {
-      DiscordRPC.discordRunCallbacks();
-    }
+    client.close();
   }
 
   @EventInvoked
   public void onGuiSwitch(GuiSwitchEvent e) {
-    switch(e.getType()) {
-      case NULL:
-        String serverIP = Sorus.getSorus().getVersion().getGame().getCurrentServerIP();
-        if(serverIP == null) {
-          this.setPresence("Playing Singleplayer");
-        } else {
-          this.setPresence("Playing " + serverIP);
-        }
-        break;
-      case MAIN_MENU:
-        this.setPresence("In Menus");
-        break;
+    if (!Sorus.getSorus().getVersion().getGame().getPlayer().isNull()) {
+      String serverIP = Sorus.getSorus().getVersion().getGame().getCurrentServerIP();
+      if (serverIP == null) {
+        this.setPresence("Playing Singleplayer");
+      } else {
+        this.setPresence("Playing " + serverIP);
+      }
+    } else {
+      this.setPresence("In Menus");
     }
   }
 
   public void setPresence(String firstLine) {
-    presence.details = firstLine;
-    DiscordRPC.discordUpdatePresence(presence);
+    builder.setDetails(firstLine);
+    client.sendRichPresence(builder.build());
   }
 }
