+++ Versions/1.8.9/src/main/java/org/sorus/oneeightnine/EntityImpl.java
@@ -50,4 +50,8 @@ public class EntityImpl implements IEntity {
         return entity.posZ;
     }
 
+    @Override
+    public boolean isNull() {
+        return entity == null;
+    }
 }
